requirements.txt
# Telegram Bot
python-telegram-bot==21.7

# Whisper transcription (local, no API key needed)
faster-whisper==1.0.3

# Environment variables
python-dotenv==1.0.1
